# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booker API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List of all users",
                "operationId": "7e5b40495c8c742505e6042c37444fa8",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "id": "1",
                                            "name": "John",
                                            "email": "john@mail.com"
                                        },
                                        {
                                            "id": "2",
                                            "name": "Tom",
                                            "email": "tom@mail.com"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Creating a new user.",
                "operationId": "a6471d589feb6b6653e755fec7bc9fbd",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            },
                            "example": {
                                "name": "Peter",
                                "email": "peter@mail.com",
                                "password": "password"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Peter"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "peter@mail.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
        "/api/v1/user/{user}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get an user",
                "operationId": "d65cc1c022ff40305e7754e67a6f3028",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id of the user",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Peter"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "peter@mail.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "operationId": "0cd4c61bc3e9ca5e2a5edff5af60fb7b",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id of the user",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Jack"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "jack@mail.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jack"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jack@mail.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete the user",
                "operationId": "6fe7c1a922e25715ba55b0c79a77a75e",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id of the user",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}